### 2024-12-17 #7598 @editorialbot
[PRE REVIEW]: calzone: A Python package for measuring calibration of probabilistic models for classification
**Submitting author:** @qiancao (<a href="http://orcid.org/0000-0002-3900-0433">Qian Cao</a>)
**Repository:** https://github.com/DIDSR/calzone/
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1.0
