### 2024-12-17 69d378e
Format the bib
