+++ paper/paper.bib
@@ -10,415 +10,274 @@
 }
 
 @article{loess_graph_austin,
-author = {Austin, Peter C. and Steyerberg, Ewout W.},
-title = {Graphical assessment of internal and external calibration of logistic regression models by using loess smoothers},
-journal = {Statistics in Medicine},
-volume = {33},
-number = {3},
-pages = {517-535},
-keywords = {logistic regression, prediction, calibration, graphical methods, prediction models},
-doi = {https://doi.org/10.1002/sim.5941},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.5941},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/sim.5941},
-abstract = {Predicting the probability of the occurrence of a binary outcome or condition is important in biomedical research. While assessing discrimination is an essential issue in developing and validating binary prediction models, less attention has been paid to methods for assessing model calibration. Calibration refers to the degree of agreement between observed and predicted probabilities and is often assessed by testing for lack-of-fit. The objective of our study was to examine the ability of graphical methods to assess the calibration of logistic regression models. We examined lack of internal calibration, which was related to misspecification of the logistic regression model, and external calibration, which was related to an overfit model or to shrinkage of the linear predictor. We conducted an extensive set of Monte Carlo simulations with a locally weighted least squares regression smoother (i.e., the loess algorithm) to examine the ability of graphical methods to assess model calibration. We found that loess-based methods were able to provide evidence of moderate departures from linearity and indicate omission of a moderately strong interaction. Misspecification of the link function was harder to detect. Visual patterns were clearer with higher sample sizes, higher incidence of the outcome, or higher discrimination. Loess-based methods were also able to identify the lack of calibration in external validation samples when an overfit regression model had been used. In conclusion, loess-based smoothing methods are adequate tools to graphically assess calibration and merit wider application. © 2013 The Authors. Statistics in Medicine published by John Wiley \& Sons, Ltd},
-year = {2014}
+  author = {Austin, Peter C. and Steyerberg, Ewout W.},
+  title = {Graphical assessment of internal and external calibration of logistic regression models by using loess smoothers},
+  journal = {Statistics in Medicine},
+  volume = {33},
+  number = {3},
+  pages = {517-535},
+  keywords = {logistic regression, prediction, calibration, graphical methods, prediction models},
+  doi = {10.1002/sim.5941},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.5941},
+  eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/sim.5941},
[…]
