+++ paper/paper.bib
@@ -281,3 +281,89 @@
 }
 
 @InProceedings{guo_calibration,
+  title = {On Calibration of Modern Neural Networks},
+  author = {Chuan Guo and Geoff Pleiss and Yu Sun and Kilian Q. Weinberger},
+  booktitle = {Proceedings of the 34th International Conference on Machine Learning},
+  pages = {1321--1330},
+  year = {2017},
+  editor = {Precup, Doina and Teh, Yee Whye},
+  volume = {70},
+  series = {Proceedings of Machine Learning Research},
+  month = {06--11 Aug},
+  publisher = {PMLR},
+  pdf = {http://proceedings.mlr.press/v70/guo17a/guo17a.pdf},
+  url = {https://proceedings.mlr.press/v70/guo17a.html},
+  abstract = {Confidence calibration – the problem of predicting probability estimates representative of the true correctness likelihood – is important for classification models in many applications. We discover that modern neural networks, unlike those from a decade ago, are poorly calibrated. Through extensive experiments, we observe that depth, width, weight decay, and Batch Normalization are important factors influencing calibration. We evaluate the performance of various post-processing calibration methods on state-of-the-art architectures with image and document classification datasets. Our analysis and experiments not only offer insights into neural network learning, but also provide a simple and straightforward recipe for practical settings: on most datasets, temperature scaling – a single-parameter variant of Platt Scaling – is surprisingly effective at calibrating predictions.}
+}
+
+@article{hl_test,
+  author = {David W. Hosmer and Stanley Lemesbow},
+  title = {Goodness of fit tests for the multiple logistic regression model},
+  journal = {Communications in Statistics - Theory and Methods},
+  volume = {9},
+  number = {10},
+  pages = {1043--1069},
+  year = {1980},
+  publisher = {Taylor \& Francis},
+  doi = {10.1080/03610928008827941},
+  url = {https://www.tandfonline.com/doi/abs/10.1080/03610928008827941},
+  eprint = {https://www.tandfonline.com/doi/pdf/10.1080/03610928008827941}
+}
+
+@article{huang_tutorial,
+  author = {Huang, Yingxiang and Li, Wentao and Macheret, Fima and Gabriel, Rodney A and Ohno-Machado, Lucila},
+  title = {A tutorial on calibration measurements and calibration models for clinical prediction models},
+  journal = {Journal of the American Medical Informatics Association},
+  volume = {27},
+  number = {4},
+  pages = {621-633},
+  year = {2020},
+  month = {02},
[…]
