+++ paper/paper.bib
@@ -44,7 +44,8 @@
   number = {158},
   pages = {209--212},
   year = {1927},
-  publisher = {Taylor \& Francis}
+  publisher = {Taylor \& Francis},
+  doi = {10.2307/2276774}
 }
 
 @article{Brocker_reldia,
@@ -78,6 +79,7 @@
 @article{beta-binomial,
   issn = {0006341X, 15410420},
   url = {http://www.jstor.org/stable/2529131},
+  doi = {10.2307/2529131},
   author = {Griffiths, D. A.},
   journal = {Biometrics},
   number = {4},
@@ -91,6 +93,7 @@
 @article{Murphy_reliability,
   issn = {00359254, 14679876},
   url = {http://www.jstor.org/stable/2346866},
+  doi = {10.2307/2346866},
   author = {Murphy, Allan H. and Winkler, Robert L.},
   journal = {Journal of the Royal Statistical Society. Series C (Applied Statistics)},
   number = {1},
@@ -153,6 +156,7 @@
 @article{gneiting2007strictly,
   title = {Strictly proper scoring rules, prediction, and estimation},
   author = {Gneiting, Tilmann and Raftery, Adrian E},
+  doi = {10.1198/016214506000001437},
   journal = {Journal of the American statistical Association},
   volume = {102},
   number = {477},
@@ -166,6 +170,7 @@
   author = {McCullagh, P. and Nelder, J. A.},
   publisher = {Chapman \& Hall / CRC},
   title = {Generalized Linear Models},
+  doi = {10.1201/9781439891148-8},
   year = {1989}
 }
 
@@ -226,6 +231,7 @@
 
 @Manual{ResourceSelection,
   title = {ResourceSelection: Resource Selection (Probability) Functions for Use-Availability Data},
+  doi = {10.32614/cran.package.resourceselection},
   author = {Lele, Subhash R. and Keim, Jonah L. and Solymos, Peter},
   year = {2024},
   note = {R package version 0.3-6},
@@ -245,6 +251,7 @@
 @InProceedings{netcal,
   author = {Küppers, Fabian and Kronenberger, Jan and Shantia, Amirhossein and Haselhoff, Anselm},
   title = {Multivariate Confidence Calibration for Object Detection},
+  doi = {10.1109/cvprw50498.2020.00171},
   booktitle = {The IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR) Workshops},
   month = {June},
   year = {2020}
@@ -338,6 +345,7 @@
 @article{spiegelhalter_z,
   title = {Probabilistic prediction in patient management and clinical trials},
   author = {Spiegelhalter, David J},
+  doi = {10.1002/sim.4780050506},
   journal = {Statistics in medicine},
   volume = {5},
   number = {5},
@@ -361,6 +369,7 @@
 @article{Walsh_overview,
   title = {Beyond discrimination: a comparison of calibration methods and clinical usefulness of predictive models of readmission risk},
   author = {Walsh, Colin G and Sharman, Kavya and Hripcsak, George},
+  doi = {10.1016/j.jbi.2017.10.008},
   journal = {Journal of biomedical informatics},
   volume = {76},
   pages = {9--18},
