### 2025-01-03 #7598 @crvernon
@jasonfan1997 I am going to run this through a review with our larger editorial board to ensure it meets our submission requirements.  I'll get back to you ASAP.

In the meantime, the paper length should be reduced to around 1000 words.

Thanks!
