### 2025-01-27 #7598 @fabian-s
hi again @adriancorrendo @schalkdaniel @giuseppec

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.

Even if not, please reply to this message so I can start looking for other potential reviewers.
