### 2025-02-11 #7598 @fabian-s
hi  @ShanaScogin @MrShoenel @brunomontezano

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.

Even if not, please reply to this message so I can start looking for other potential reviewers.
