### 2025-02-11 #7598 @adriancorrendo
Hi, sorry the delay, I had a problem with my notifications settings. Anyways, I won't be able to review this as Python is not part of my expertise.

Best,
Adrian
