### 2025-02-11 #7598 @fabian-s
hi @hbaniecki

would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.

Even if not, it would be kind of you reply to this message so I can start looking for other potential reviewers.
