### 2025-03-03 #7598 @fabian-s
👋 @arunmano121, @andrewheiss, [@esadr](https://reviewers.joss.theoj.org/reviewers/736)

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Even if not, it would be kind of you reply to this message so I can start looking for other potential reviewers.
