### 2025-03-05 f866ea9
Add example to README.md
