### 2025-03-06 #7598 @andrewheiss
Oh, apologies for the delay here! I also don't have time/bandwidth to handle this review :(
