### 2025-03-21 3dab272
Remove GUI install since it is optional
