+++ pyproject.toml
@@ -32,11 +32,10 @@ Issues = "https://github.com/DIDSR/calzone/issues"
 
 [project.scripts]
 cal_metrics = "cal_metrics:main"
-GUI_cal_metrics = "GUI_cal_metrics:main"
 
 [tool.setuptools]
 packages = ["calzone"]
-py-modules = ["cal_metrics", "GUI_cal_metrics"]
+py-modules = ["cal_metrics"]
 
 [tool.pytest.ini_options]
 addopts = "-v --cov=calzone"
