+++ calzone_demo.py
@@ -1,0 +1,227 @@
+"""
+
+My workflow so far:
+
+conda create -n calzone-gui # new clean environment
+
+conda activate it
+install pip
+install all the dependencies with pip
+install pyinstaller
+install nicegui
+install pywin32
+
+nicegui-pack --onefile --name "Calzone" calzone_demo.py
+
+check for .exe file in /dist
+
+see: https://nicegui.io/documentation/section_configuration_deployment#package_for_installation
+
+"""
+
+import argparse
+import subprocess
+from nicegui import ui, app, events
+from typing import Optional
+from pathlib import Path
+import platform
+
+from local_file_picker import local_file_picker
+
+from nicegui import native
+
+def run_program():
+    clear_cache()
+    output_area.value = ""  # Clear the output area
+    plot_image.clear()
+    plot_image.set_source(None)  # Set the image source to None
+    ui.update(plot_image)
+    csv_file = csv_file_input.value
+    save_metrics = save_metrics_input.value
+    save_plot = save_plot_input.value
+
+    selected_metrics = [metric for metric, checkbox in metrics_checkboxes.items()
+                        if checkbox.value]
+
+    if not selected_metrics or not csv_file:
+        ui.notify("Error: Please select at least one metric and provide a CSV file.",
+                  type="error")
+        return
+
+    metric_arg = "all" if not selected_metrics else ",".join(selected_metrics)
+
+    if perform_bootstrap_checkbox.value and int(n_bootstrap_input.value) == 0:
+        ui.notify("Error: Number of bootstrap samples must be greater than zero when performing bootstrap.",
+                  type="error")
+        return
+
+    args = [
+        "--csv_file", str(csv_file),
+        "--metrics", str(metric_arg),
+        "--class_to_calculate", str(class_to_calculate_input.value),
+        "--num_bins", str(num_bins_input.value),
+        "--plot_bins", str(plot_bins_input.value),
+    ]
+
+    if save_metrics_input.value:
+        args.append("--save_metrics")
+        args.append(str(save_metrics))
+
+    if perform_bootstrap_checkbox.value:
+        args.append("--n_bootstrap")
+        args.append(str(int(n_bootstrap_input.value)))
+        args.append("--bootstrap_ci")
+        args.append(str(bootstrap_ci_input.value))
+
+    if prevalence_adjustment_checkbox.value:
+        args.append("--prevalence_adjustment")
+
+    if plot_checkbox.value:
+        if not save_plot:
+            ui.notify("Error: Please provide a path to save the plot.", type="error")
+            return
+        args.append("--plot")
+        args.append("--save_plot")
+        args.append(str(save_plot))
+    if verbose_checkbox.value:
+        args.append("--verbose")
+    if topclass_checkbox.value:
+        args.append("--topclass")
+    if hl_test_validation_checkbox.value:
+        if not any(metric in selected_metrics for metric in ["HL-H", "HL-C"]):
+            ui.notify("Error: HL test validation requires either HL-H or HL-C metric to be selected.", type="error")
+            return
+        args.append("--hl_test_validation")
+
+    command = ["cal_metrics"] + args
+    print("Running command:", " ".join(command))
+
+    process = subprocess.Popen(command, stdout=subprocess.PIPE,
+                               stderr=subprocess.PIPE, universal_newlines=True)
+    output, error = process.communicate()
+
+    output_area.value = output
+    if error:
+        output_area.value += "\nErrors:\n" + error
+
+    if plot_checkbox.value:
+        display_plot(save_plot)
+
+
+def display_plot(plot_path):
+    plot_image.clear()
+    plot_image.set_source(plot_path)
+    ui.update(plot_image)
+
+
+def clear_cache():
+    ui.run_javascript('''
+        function clearCache() {
[…]
