+++ calzone_demo.py
@@ -26,9 +26,279 @@ from typing import Optional
 from pathlib import Path
 import platform
 
-from local_file_picker import local_file_picker
+#from local_file_picker import local_file_picker
 
 from nicegui import native
+
+### Instead of running command line, run it directly
+from argparse import Namespace
+import argparse
+import numpy as np
+from calzone.metrics import CalibrationMetrics, get_CI
+from calzone.utils import *
+from calzone.vis import plot_reliability_diagram
+import io
+import sys
+from contextlib import redirect_stdout, redirect_stderr
+
+def perform_calculation(probs, labels, args, suffix=""):
+    """
+    Calculate calibration metrics and visualize reliability diagram.
+
+    Args:
+        probs (numpy.ndarray): Predicted probabilities for each class.
+        labels (numpy.ndarray): True labels.
+        args (argparse.Namespace): Command-line arguments.
+        suffix (str, optional): Suffix for output files. Defaults to "".
+
+    Returns:
+        numpy.ndarray: Calculated metrics and confidence intervals (if bootstrapping is used).
+    """
+    cal_metrics = CalibrationMetrics(
+        class_to_calculate=args.class_to_calculate, num_bins=args.num_bins
+    )
+    if args.hl_test_validation:
+        df = args.num_bins
+    else:
+        df = args.num_bins - 2
+    metrics_to_calculate = args.metrics.split(",") if args.metrics else ["all"]
+    if metrics_to_calculate == ["all"]:
+        metrics_to_calculate = "all"
+    result = cal_metrics.calculate_metrics(
+        y_true=labels,
+        y_proba=probs,
+        metrics=metrics_to_calculate,
+        perform_pervalance_adjustment=args.prevalence_adjustment,
+        df = df
+    )
+
+    keys = list(result.keys())
+    result = np.array(list(result.values())).reshape(1, -1)
+
+    if args.n_bootstrap > 0:
+        bootstrap_results = cal_metrics.bootstrap(
+            y_true=labels,
+            y_proba=probs,
+            n_samples=args.n_bootstrap,
+            metrics=metrics_to_calculate,
+            perform_pervalance_adjustment=args.prevalence_adjustment,
+            df = df
+        )
+        CI = get_CI(bootstrap_results)
+        result = np.vstack((result, np.array(list(CI.values())).T))
+
+    if args.verbose:
+        print_metrics(result, keys, args.n_bootstrap, suffix)
+
+    if args.save_metrics:
+        save_metrics_to_csv(result, keys, args.save_metrics, suffix)
+
+    if args.plot:
+        plot_reliability(labels, probs, args, suffix)
+    return result
+
+
+def print_metrics(result, keys, n_bootstrap, suffix):
+    """
+    Print calculated metrics.
+
+    Args:
+        result (numpy.ndarray): Calculated metrics and confidence intervals.
+        keys (list): Names of the calculated metrics.
+        n_bootstrap (int): Number of bootstrap samples.
+        suffix (str): Suffix for output files.
+    """
+    if n_bootstrap > 0:
+        print_header = (
+            "Metrics with bootstrap confidence intervals:"
+            if suffix == ""
+            else f"Metrics for {suffix} with bootstrap confidence intervals:"
+        )
+        print(print_header)
+        for i, num in enumerate(keys):
+            print(
+                f"{num}: {np.format_float_positional(result[0][i], 3)}",
+                f"({np.format_float_positional(result[1][i], 3)}, "
+                f"{np.format_float_positional(result[2][i], 3)})",
+            )
+    else:
+        print_header = "Metrics:" if suffix == "" else f"Metrics for subgroup {suffix}:"
+        print(print_header)
+        for i, num in enumerate(keys):
+            print(f"{num}: {np.format_float_positional(result[0][i], 3)}")
+
+
[…]
