+++ docs/build/latex/calzone.tex
@@ -65,7 +65,7 @@
 
 
 \title{calzone}
-\date{Dec 04, 2024}
+\date{Apr 04, 2025}
 \release{develop}
 \author{Kwok Lung "Jason" Fan, Qian Cao}
 \newcommand{\sphinxlogo}{\vbox{}}
@@ -85,9 +85,6 @@
 \pagestyle{normal}
 \phantomsection\label{\detokenize{index::doc}}
 
-
-\sphinxAtStartPar
-calzone is a Python package for calculation of various calibration metrics. This work is credited to Kwok Lung (Jason) Fan and Qian Cao.
 
 \sphinxAtStartPar
 The calzone package provides a suite of tools for assessing and improving the calibration of machine learning models, particularly for binary classification tasks. It offers calibration metrics and visualization tools for displaying reliability diagrams.  Please read the summary and guide section in this documentation first before using the package.
@@ -139,23 +136,20 @@ This tutorial provides a quick start to the calzone package, including a discuss
 \section{Installation}
 \label{\detokenize{notebooks/quickstart:Installation}}
 \sphinxAtStartPar
-Calzone dependencies are numpy, scipy, statsmodels and matplotlib. If you are experienced developer, you probably already have numpy, scipy and matplolib installed. If you don’t have a package install, you can install them with conda
+We recommend that you install \sphinxcode{\sphinxupquote{calzone}} from pip:
 
 \begin{sphinxuseclass}{nbinput}
 \begin{sphinxuseclass}{nblast}
 {
 \begin{sphinxVerbatim}[commandchars=\\\{\}]
-\llap{\color{nbsphinxin}[ ]:\,\hspace{\fboxrule}\hspace{\fboxsep}}\PYG{n}{conda} \PYG{n}{install} \PYG{n}{numpy}
-\PYG{n}{conda} \PYG{n}{install} \PYG{n}{scipy}
-\PYG{n}{conda} \PYG{n}{install} \PYG{n}{matplotlib}
-\PYG{n}{conda} \PYG{n}{install} \PYG{n}{statsmodels}
+\llap{\color{nbsphinxin}[ ]:\,\hspace{\fboxrule}\hspace{\fboxsep}}\PYG{n}{pip} \PYG{n}{install} \PYG{n}{calzone}\PYG{o}{\PYGZhy{}}\PYG{n}{tool}
 \end{sphinxVerbatim}
 }
 
 \end{sphinxuseclass}
 \end{sphinxuseclass}
 \sphinxAtStartPar
-Alternatively, you can install the dependencies with pip
+Alternatively, you can install the package and its dependencies manually. Calzone has very few dependencies: NumPy, SciPy, Statsmodels and Matplotlib:
 
 \begin{sphinxuseclass}{nbinput}
 \begin{sphinxuseclass}{nblast}
@@ -171,7 +165,7 @@ Alternatively, you can install the dependencies with pip
 \end{sphinxuseclass}
 \end{sphinxuseclass}
 \sphinxAtStartPar
-Then, you can install the calzone package from github if you only want to use the calculator inside you python script:
+Then, the package can be installed directly from the GitHub repository:
 
 \begin{sphinxuseclass}{nbinput}
 \begin{sphinxuseclass}{nblast}
[…]
