### 2025-04-07 8b06b3b
Adding logo
