+++ gui/calzoneGUI.py
@@ -1,6 +1,6 @@
 """
 
-My workflow so far:
+Creating the exe for calzone GUI
 
 conda create -n calzone-gui # new clean environment
 
@@ -30,7 +30,6 @@ import platform
 
 from nicegui import native
 
-### Instead of running command line, run it directly
 from argparse import Namespace
 import argparse
 import numpy as np
@@ -413,9 +412,13 @@ async def pick_file() -> None:
         update_csv_file_input(result[0])
 
 
+
+
 with ui.row().classes('w-full justify-center'):
     with ui.column().classes('w-1/3 p-4'):
-        ui.label('calzone GUI').classes('text-h4')
+        with ui.row().classes('items-center gap-4'):
+            ui.image('../logo.png').classes('w-24 h-24')  # Made logo bigger
+            ui.label('calzone GUI').classes('text-h4')
         csv_file_input = ui.input(label='CSV File',
                                   placeholder='Enter absolute file path').classes('w-full')
         ui.button('choose file', on_click=pick_file)
