### 2025-04-07 1adabd4
tested /gui and added a minimal build script
