### 2025-04-07 b928ef5
tested /gui and added a minimal build script
