+++ gui/calzoneGUI.py
@@ -36,6 +36,7 @@ import numpy as np
 from calzone.metrics import CalibrationMetrics, get_CI
 from calzone.utils import *
 from calzone.vis import plot_reliability_diagram
+from matplotlib import pyplot as plt
 import io
 import sys
 from contextlib import redirect_stdout, redirect_stderr
@@ -93,8 +94,10 @@ def perform_calculation(probs, labels, args, suffix=""):
         save_metrics_to_csv(result, keys, args.save_metrics, suffix)
 
     if args.plot:
-        plot_reliability(labels, probs, args, suffix)
-    return result
+        fig = plot_reliability(labels, probs, args, suffix)
+    else:
+        fig = None
+    return result, fig
 
 
 def print_metrics(result, keys, n_bootstrap, suffix):
@@ -176,11 +179,11 @@ def plot_reliability(labels, probs, args, suffix):
             diagram_filename = pathwithoutextension + "_" + suffix + ".csv"
         else:
             diagram_filename = None
-
-    valid_image_formats = ['.png', '.jpg', '.jpeg', '.gif', '.bmp', '.tiff', '.pdf']
-    if not any(filename.lower().endswith(fmt) for fmt in valid_image_formats):
-        print(filename)
-        raise ValueError("Invalid file format. Please provide a valid image format.")
+    if args.save_plot is not None:
+        valid_image_formats = ['.png', '.jpg', '.jpeg', '.gif', '.bmp', '.tiff', '.pdf']
+        if not any(filename.lower().endswith(fmt) for fmt in valid_image_formats):
+            print(filename)
+            raise ValueError("Invalid file format. Please provide a valid image format.")
 
     reliability, confidence, bin_edge, bin_count = reliability_diagram(
         y_true=labels,
@@ -189,7 +192,7 @@ def plot_reliability(labels, probs, args, suffix):
         class_to_plot=args.class_to_calculate,
         save_path=diagram_filename,
     )
-    plot_reliability_diagram(
+    fig = plot_reliability_diagram(
         reliability,
         confidence,
         bin_count,
@@ -198,7 +201,9 @@ def plot_reliability(labels, probs, args, suffix):
         error_bar=True,
         return_fig=True
     )
-    print("Plot saved to", filename)
+    if args.save_plot is not None:
+        print("Plot saved to", filename)
+    return fig
 
 
 def run_calibration(args):
@@ -208,11 +213,11 @@ def run_calibration(args):
         loader = loader.transform_topclass()
 
     if not loader.have_subgroup:
-        perform_calculation(
+        _ ,fig = perform_calculation(
             probs=loader.probs, labels=loader.labels, args=args, suffix=""
         )
     else:
-        perform_calculation(
+        _ ,fig = perform_calculation(
             probs=loader.probs, labels=loader.labels, args=args, suffix=""
         )
         for i, subgroup_column in enumerate(loader.subgroup_indices):
@@ -225,6 +230,7 @@ def run_calibration(args):
                     args=args,
                     suffix=f"subgroup_{i+1}_group_{subgroup_class}",
                 )
+    return 1, fig
 
 
 class local_file_picker(ui.dialog):
@@ -306,9 +312,6 @@ class local_file_picker(ui.dialog):
 def run_program():
     #clear_cache()
     output_area.value = ""  # Clear the output area
-    plot_image.clear()
-    plot_image.set_source(None)  # Set the image source to None
-    ui.update(plot_image)
     # Get values from UI elements
     csv_file = csv_file_input.value
     save_metrics = save_metrics_input.value
@@ -337,11 +340,11 @@ def run_program():
                     type="error")
             return
     
-    if plot_checkbox.value:
-        if save_plot == "":
[…]
