+++ docs/build/latex/calzone.tex
@@ -65,7 +65,7 @@
 
 
 \title{calzone}
-\date{Apr 04, 2025}
+\date{Apr 07, 2025}
 \release{develop}
 \author{Kwok Lung "Jason" Fan, Qian Cao}
 \newcommand{\sphinxlogo}{\vbox{}}
