### 2025-04-07 821e975
Add matplotlib backend
