+++ gui/calzoneGUI.py
@@ -423,7 +423,7 @@ with ui.row().classes('w-full justify-center'):
     with ui.column().classes('w-1/3 p-4'):
         with ui.row().classes('items-center gap-4'):
             ui.image('../logo.png').classes('w-24 h-24')  # Made logo bigger
-            ui.label('calzone GUI').classes('text-h4')
+            ui.label('Calzone GUI (experimental)').classes('text-h4')
         csv_file_input = ui.input(label='CSV File',
                                   placeholder='Enter absolute file path').classes('w-full')
         ui.button('choose file', on_click=pick_file)
