### 2025-04-11 #8026 @fabian-s
@patRyserWelch8 I apologize for the confusion -- when I asked you to review this, I had overlooked that I had *already* recruited two other reviewers, so that I would not have needed to burden you as well.
Unfortunately, there is no way to formally retract a reviewer invitation using the JOSS editorial bot.
You can simply ignore any messages coming your way from this GitHub thread (or use the unsubcsribe link in the e-mail or go to https://github.com/openjournals/joss-reviews/issues/8026 and click on "unsubscribe" in the notifications panel (top right)) -- not sure why Github is still forwarding you these as e-mails.
Of course, if you have already started your review, you're also very welcome to simply share your feedback for the authors informally in this thread.
