### 2025-04-14 #8026 @patRyserWelch8
That is no problem.  Happy to help another time.

Pat

From: Fabian Scheipl ***@***.***>
Date: Friday, 11 April 2025 at 10:02
Cc: Pat.Ryser-Welch ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: calzone: A Python package for measuring calibration of probabilistic models for classification (Issue #8026)

External email: if the sender or content looks suspicious, please click the Report Message icon, or forward it to report-phishing

@patRyserWelch8<https://github.com/patRyserWelch8> I apologize for the confusion -- when I asked you to review this, I had overlooked that I had already recruited two other reviewers, so that I would not have needed to burden you as well.
Unfortunately, there is no way to formally retract a reviewer invitation using the JOSS editorial bot.
You can simply ignore any messages coming your way from this GitHub thread (or use the unsubcsribe link in the e-mail or go to #8026<https://github.com/openjournals/joss-reviews/issues/8026> and click on "unsubscribe" in the notifications panel (top right)) -- not sure why Github is still forwarding you these as e-mails.
Of course, if you have already started your review, you're also very welcome to simply share your feedback for the authors informally in this thread.
