### 2025-04-28 500b425
Add score and ROC plot
