+++ gui/calzoneGUI.py
@@ -20,9 +20,7 @@ see: https://nicegui.io/documentation/section_configuration_deployment#package_f
 """
 
 import io
-import sys
-import argparse
-import subprocess
+import os
 from nicegui import ui, app, events
 from typing import Optional
 from pathlib import Path
@@ -35,7 +33,6 @@ from nicegui import native
 from argparse import Namespace
 import argparse
 import numpy as np
-
 from calzone.metrics import CalibrationMetrics, get_CI
 from calzone.utils import *
 from calzone.vis import plot_reliability_diagram
@@ -43,6 +40,7 @@ import io
 import sys
 from contextlib import redirect_stdout, redirect_stderr
 import matplotlib
+import matplotlib.pyplot as plt
 matplotlib.use("svg")
 
 def perform_calculation(probs, labels, args, suffix=""):
@@ -209,6 +207,37 @@ def plot_reliability(labels, probs, args, suffix):
         print("Plot saved to", filename)
     return fig
 
+def plot_hist_roc(labels, probs, args):
+    if args.class_to_calculate is None:
+        class_to_calculate = 1
+    else:
+        class_to_calculate = args.class_to_calculate
+    # Calculate ROC curve
+    # Create figure with two subplots
+    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(12, 5))
+    labels = labels.flatten()
+    # Plot histogram
+    pos_scores = probs[labels == class_to_calculate, class_to_calculate]
+    neg_scores = probs[labels != class_to_calculate, class_to_calculate]
+    ax1.hist(pos_scores, bins=20, alpha=0.5, label='Class-of-interest', density=False)
+    ax1.hist(neg_scores, bins=20, alpha=0.5, label='Rest', density=False)
+    ax1.set_xlabel('Prediction score')
+    ax1.set_ylabel('Density')
+    ax1.legend()
+    ax1.set_title('Score distributions')
+    
+    # Plot ROC curve
+    fpr, tpr, roc_auc = make_roc_curve(labels, probs, class_to_plot=class_to_calculate)
+    ax2.plot(fpr, tpr, label=f'ROC curve (AUC = {roc_auc:.3f})')
+    ax2.plot([0, 1], [0, 1], 'k--')
+    ax2.set_xlabel('False Positive Rate')
+    ax2.set_ylabel('True Positive Rate') 
+    ax2.legend()
+    ax2.set_title('ROC Curve')
+    
+    plt.tight_layout()
+    return fig
+
 
 def run_calibration(args):
     loader = data_loader(args.csv_file)
@@ -220,6 +249,8 @@ def run_calibration(args):
         _ ,fig = perform_calculation(
             probs=loader.probs, labels=loader.labels, args=args, suffix=""
         )
+        if args.plot_hist_roc:
+            fig2 = plot_hist_roc(loader.labels, loader.probs, args)
     else:
         _ ,fig = perform_calculation(
             probs=loader.probs, labels=loader.labels, args=args, suffix=""
@@ -234,7 +265,9 @@ def run_calibration(args):
                     args=args,
                     suffix=f"subgroup_{i+1}_group_{subgroup_class}",
                 )
-    return 1, fig
+                if args.plot_hist_roc:
+                    fig2 = plot_hist_roc(label, proba, args)
+    return 1, fig, fig2
 
 
 class local_file_picker(ui.dialog):
@@ -361,6 +394,7 @@ def run_program():
         bootstrap_ci=float(bootstrap_ci_input.value) if perform_bootstrap_checkbox.value else None,
         prevalence_adjustment=bool(prevalence_adjustment_checkbox.value),
         plot=bool(plot_checkbox.value),
+        plot_hist_roc=bool(plot_hist_roc_checkbox.value),
[…]
