### 2025-04-28 931f133
Fix subgroup bug in calzoneGUI
