### 2025-04-29 5b1f04a
Fixing and adding logo
