### 2025-05-07 a977d6e
Update README.md
mentioned independence of the testing dataset in the readme file
