### 2025-06-28 DIDSR/calzone#5 @pescap
JOSS Review #2
This issue is related to [openjournals/joss-reviews/issues/8026](https://github.com/openjournals/joss-reviews/issues/8026).

Hi, very good work on such an important subject!

# Paper

- Pag.1, l.18: What is Se/Sp? Add ref for AUC, Se/Sp.
- In the paper you state “provides functions to compute various calibration metrics.” Here I would add more guidance and general context before listing them one by one — to help the reader understand that some metrics do X, others do Y, and so on.

Try to show in which ways you are extending and adding specialized features to the generic `scikit-learn` functionality. When should one use `calzone`? This could go in the statement of interest. Because again, I feel like one does not need to be an expert in calibration to need to go beyond `scikit-learn` and use `calzone`.

For an introduction to this subject, `scikit-learn` entry on [calibration](https://scikit-learn.org/stable/modules/calibration.html) good as an introduction. Maybe actually, the reader could start with that, and then get to why calzone, and the specificities you add.

This boundary between: go with `scikit-learn` v/s use `calzone` is key to me. You mention it in the statement of need but I would be more specific.

Also, mention other libraries and how you present a novelty and differentiation, e.g.:
- [ml-calibration](https://github.com/apple/ml-calibration)
- [pycaleva](https://github.com/MartinWeigl/pycaleva)

# Doc

- "calzone aims to assess whether a model achieves moderate calibration, meaning whether [...]". The notation for this equation is different from the one in *Statement of need*. Maybe use the same consistent notation across the docs and JOSS paper.

In [metrics_summary](https://calzone-docs.readthedocs.io/en/latest/notebooks/metrics_summary.html):
- The table is interesting, but with the ReadTheDocs default zoom I cannot read it well, it appears too small. Maybe turn the table into a Markdown table.

In general, for the docs, I would be more "guiding".

- Maybe add references or small explanations for a non-expert audience. For example:

```
To accurately assess the calibration of machine learning models, it is essential to have a comprehensive and
representative dataset with sufficient coverage of the prediction space.
The calibration metric is not meaningful if the dataset is not representative of the true intended population.
```

and

```
If you observe general over- or under-estimation of probabilities for a given class,
consider applying a prevalence adjustment to determine if it’s solely due to a prevalence
shift. After prevalence adjustment, plot the reliability diagrams again and examine
the results of calibration metrics.
```

"Representative dataset with sufficient coverage": explain a bit more.
I do not master the definitions like "prevalence", "prevalence adjustment" (which type, what).
Or maybe refer to the dedicated section in the docs.

Check typos: *representative*, *alternatively*.

[reliability_diagram](https://calzone-docs.readthedocs.io/en/latest/notebooks/reliability_diagram.html): check typos. Give more details. For example, detail what is written as comments:

```
### We will artificially drop the prevalence of class 1
# The top-class reliability diagram will still look good
# But the class-1 reliability diagram will be very bad
```

Give a detail of the different kinds of methods here. Because afterwards, in the next sections, you detail the different tests that have diverse usages. I would summarize the broad usage of all at the beginning and also in the paper. My impression when I started with the dict was that there were a lot of things measured, but not a clear view of the whole. Maybe be more consistent in explaining the example data too.

[Validating metrics](https://calzone-docs.readthedocs.io/en/latest/notebooks/validation.html): add these tests (or some of them) to the unit tests?

# Curiosity

What about unbalanced datasets? Maybe elaborate on this.
