### 2025-06-30 9e8aac8
update paper to address reviewer comments
