+++ paper/paper.bib
@@ -341,7 +341,36 @@
   number = {7},
   year = {2019}
 }
+@misc{pycaleva,
+  author       = {Martin Weigl, Mark Aron Szulyovszky},
+  title        = {pycaleva},
+  year         = {2022},
+  howpublished = {\url{https://github.com/MartinWeigl/pycaleva}},
+  note         = {GitHub repository},
+}
 
+@inproceedings{ml-calibration,
+      title={Smooth {ECE}: Principled Reliability Diagrams via Kernel Smoothing},
+      author={B{\l}asiok, Jaros{\l}aw and Nakkiran, Preetum},
+      booktitle={The Twelfth International Conference on Learning Representations},
+      year={2024},
+      url={https://openreview.net/forum?id=XwiA1nDahv}
+}
+
+@book{statistical_learning,
+  added-at = {2008-05-16T16:17:42.000+0200},
+  address = {New York, NY, USA},
+  author = {Hastie, Trevor and Tibshirani, Robert and Friedman, Jerome},
+  biburl = {https://www.bibsonomy.org/bibtex/2f58afc5c9793fcc8ad8389824e57984c/sb3000},
+  interhash = {d585aea274f2b9b228fc1629bc273644},
+  intrahash = {f58afc5c9793fcc8ad8389824e57984c},
+  keywords = {ml statistics},
+  publisher = {Springer New York Inc.},
+  series = {Springer Series in Statistics},
+  timestamp = {2008-05-16T16:17:43.000+0200},
+  title = {The Elements of Statistical Learning},
+  year = 2001
+}
 @article{spiegelhalter_z,
   title = {Probabilistic prediction in patient management and clinical trials},
   author = {Spiegelhalter, David J},
