+++ docs/build/html/_sources/notebooks/metrics_summary.ipynb.txt
@@ -11,18 +11,85 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 1,
-   "metadata": {
-    "nbsphinx": "hidden"
-   },
+   "execution_count": 3,
+     "nbsphinx": {
+    "hide_input": true
+  },
    "outputs": [
     {
-     "name": "stderr",
-     "output_type": "stream",
-     "text": [
-      "/home/kwoklung.fan/anaconda3/envs/uq/lib/python3.12/site-packages/dataframe_image/converter/matplotlib_table.py:147: FutureWarning: The behavior of this method will change in future versions. Use specific 'len(elem)' or 'elem is not None' test instead.\n",
-      "  if not thead and not tbody:\n"
-     ]
+     "data": {
+      "text/html": [
+       "<style type=\"text/css\">\n",
+       "#T_696e7 th {\n",
+       "  text-align: center;\n",
+       "}\n",
+       "#T_696e7_row0_col0, #T_696e7_row0_col1, #T_696e7_row0_col2, #T_696e7_row0_col3, #T_696e7_row0_col4, #T_696e7_row1_col0, #T_696e7_row1_col1, #T_696e7_row1_col2, #T_696e7_row1_col3, #T_696e7_row1_col4, #T_696e7_row2_col0, #T_696e7_row2_col1, #T_696e7_row2_col2, #T_696e7_row2_col3, #T_696e7_row2_col4, #T_696e7_row3_col0, #T_696e7_row3_col1, #T_696e7_row3_col2, #T_696e7_row3_col3, #T_696e7_row3_col4, #T_696e7_row4_col0, #T_696e7_row4_col1, #T_696e7_row4_col2, #T_696e7_row4_col3, #T_696e7_row4_col4, #T_696e7_row5_col0, #T_696e7_row5_col1, #T_696e7_row5_col2, #T_696e7_row5_col3, #T_696e7_row5_col4 {\n",
+       "  text-align: left;\n",
+       "  white-space: pre-wrap;\n",
+       "}\n",
+       "</style>\n",
+       "<table id=\"T_696e7\">\n",
+       "  <thead>\n",
+       "    <tr>\n",
+       "      <th id=\"T_696e7_level0_col0\" class=\"col_heading level0 col0\" >Metrics</th>\n",
+       "      <th id=\"T_696e7_level0_col1\" class=\"col_heading level0 col1\" >Description</th>\n",
+       "      <th id=\"T_696e7_level0_col2\" class=\"col_heading level0 col2\" >Pros</th>\n",
+       "      <th id=\"T_696e7_level0_col3\" class=\"col_heading level0 col3\" >Cons</th>\n",
+       "      <th id=\"T_696e7_level0_col4\" class=\"col_heading level0 col4\" >Meaning</th>\n",
+       "    </tr>\n",
+       "  </thead>\n",
+       "  <tbody>\n",
+       "    <tr>\n",
+       "      <td id=\"T_696e7_row0_col0\" class=\"data row0 col0\" >Expected calibration error<br>(ECE)</td>\n",
+       "      <td id=\"T_696e7_row0_col1\" class=\"data row0 col1\" ><div>Using binned reliability diagram<br>(equal-width or equal-count binning),<br>sum of absolute difference, weighted by bin count.</div></td>\n",
+       "      <td id=\"T_696e7_row0_col2\" class=\"data row0 col2\" ><div>• Intuitive<br>• Easy to calculate</div></td>\n",
+       "      <td id=\"T_696e7_row0_col3\" class=\"data row0 col3\" ><div>• Depend on binning <br>• Depend on class-by-class/top-class</div></td>\n",
[…]
