+++ tests/test_metrics.py
@@ -6,6 +6,8 @@ The main test function generates synthetic data and tests the CalibrationMetrics
 
 import numpy as np
 from calzone.metrics import CalibrationMetrics
+from scipy.stats import beta
+
 
 def run_test_metrics():
     """
@@ -26,12 +28,16 @@ def run_test_metrics():
     # Generate sample data
     np.random.seed(123)
     n_samples = 1000
-    y_true = np.random.binomial(1, 0.3, n_samples)
     y_proba = np.zeros((n_samples,2))
-    y_proba[:, 1] = np.random.beta(0.5, 0.5, (n_samples, ))
-    y_proba[:, 0] = 1 - y_proba[:, 1]  # Ensure probabilities sum to 1
+    class1_proba = beta.rvs(0.5, 0.5, size=n_samples)
+    y_proba[:, 1] = class1_proba
+    y_proba[:, 0] = 1 - class1_proba  # Ensure
+    y_true = np.random.binomial(1, p=class1_proba)
 
     # Test CalibrationMetrics class
     metrics = CalibrationMetrics(class_to_calculate=1)
     results = metrics.calculate_metrics(y_true, y_proba, metrics='all')
     assert isinstance(results, dict)
+    ### Determintistic results given the random seed
+    assert results['ECE-H'] < 0.05
+    assert results['HL-H p-value'] > 0.05
