### 2025-07-09 DIDSR/calzone#5 @fabian-s
@qiancao while we wait for the other reviewer, please start adressing these points
