### 2025-07-21 #8026 @fabian-s
@qiancao please adress the issues raised by @pescap while we wait for @MrShoenel's review
