### 2025-07-23 DIDSR/calzone#6 @MrShoenel
Joss Review #1
This issue is related to <https://github.com/openjournals/joss-reviews/issues/8026>.

### Code and Documentation

* The code is well organized and adheres to modern standards.
* Functions are well documented (e.g., Args, Returns, Note, etc.). However, type and return hints could be added.
    * As a sub-note, it would be great if `CalibrationMetrics::calculate_metrics(..)` were to return an instance of some result-class or at least a named tuple.
* The documentation is, as far as I can tell, complete. I like the extensive usage of example code, plots, and relevant references.
* The biggest problem that I do see is the lack of tests. There is one test currently that essentially only checks whether "it works".
    * Ideally, aim for a high coverage of *relevant* tests. For `calzone`, that would mean that you use some example data for which you know the exact metrics' values.
    * Also, it would be important to test border cases (e.g., to few data, too extreme data, all data are the same, wrong format, etc.) and make sure the software behaves appropriately (raise warnings/errors that are helpful).
    * It would perhaps be an idea worth considering to put test files along the code files, e.g., you have `metrics.py` that could be accompanied by `metrics_test.py`.
* I have checked out the GUI (with the provided dataset) and it works reasonably well.

### Paper

* The Statement of Need is good.
* I like the Software Description; the short overview of Input data, Reliability Diagram, and the Calibration metrics is straightforward.
    * About the `Metrics` class mentioned here: What is the cost of always calculating all metrics? If cheap, why not always calculate all?
* The paper mentions that the methods were verified and that verification code exists. However, I did not find sufficient testing (see above).
    * Also, perhaps it would be a good idea to re-use the existing test data and -results from other, known libraries (like those mentioned). Then, the results should be within numeric round-off errors (well below the currently claimed 1%).
* Minor points: add a comma before "and" if you enumerate (e.g., in abstract where enumerate the supported metrics).
* The references look good; some could be improved:
    * `nixon_ace`: see <https://dblp2.uni-trier.de/rec/conf/cvpr/NixonDZJT19.html?view=bibtex>
    * `ml-calibration`: <https://dblp2.uni-trier.de/rec/conf/iclr/BlasiokN24.html?view=bibtex>
    * `netcal`: <https://dblp2.uni-trier.de/rec/conf/cvpr/KuppersKSH20.html?view=bibtex>
    * In general: ArXiv pre-prints now come with a DOI that could be added to, e.g., `taquet2022mapie`, `uncertaintyToolbox`, etc.
    * Some `@book` entries have no ISBNs.

### Checklist

Boxes I did not tick off:

- [ ] **Contribution and authorship**: There seem to be two main authors. The paper itself states 8 authors. I do not have the means to verify if all of these authors contributed in a way that would warrant co-authorship. Co-authorship usually warrants for substantial contributions to the writing of the paper.
- [ ] **Community guidelines**: The guidelines for contributing are missing. This could be handled by adding a `CONTRIBUTING.md` to the repo.
