### 2025-07-23 0217df3
minor edits to paper/paper.bib and paper/paper.md
