+++ paper/paper.bib
@@ -136,6 +136,7 @@
   title = {Applied logistic regression},
   author = {Hosmer Jr, David W and Lemeshow, Stanley and Sturdivant, Rodney X},
   year = {2013},
+  isbn = {9781118548387},
   publisher = {John Wiley \& Sons}
 }
 
@@ -170,6 +171,7 @@
   author = {McCullagh, P. and Nelder, J. A.},
   publisher = {Chapman \& Hall / CRC},
   title = {Generalized Linear Models},
+  isbn = {978-0412317606},
   doi = {10.1201/9781439891148-8},
   year = {1989}
 }
@@ -219,6 +221,7 @@
   title = {MAPIE: an open-source library for distribution-free uncertainty quantification},
   author = {Taquet, Vianney and Blot, Vincent and Morzadec, Thomas and Lacombe, Louis and Brunel, Nicolas},
   journal = {arXiv preprint arXiv:2207.12274},
+  doi = {10.48550/arXiv.2207.12274},
   year = {2022}
 }
 
@@ -226,6 +229,7 @@
   title = {Uncertainty Toolbox: an Open-Source Library for Assessing, Visualizing, and Improving Uncertainty Quantification},
   author = {Chung, Youngseog and Char, Ian and Guo, Han and Schneider, Jeff and Neiswanger, Willie},
   journal = {arXiv preprint arXiv:2109.10254},
+  doi = {10.48550/arXiv.2109.10254},
   year = {2021}
 }
 
@@ -249,12 +253,21 @@
 }
 
 @InProceedings{netcal,
-  author = {Küppers, Fabian and Kronenberger, Jan and Shantia, Amirhossein and Haselhoff, Anselm},
-  title = {Multivariate Confidence Calibration for Object Detection},
-  doi = {10.1109/cvprw50498.2020.00171},
-  booktitle = {The IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR) Workshops},
-  month = {June},
-  year = {2020}
+  author       = {Fabian K{\"{u}}ppers and
+                  Jan Kronenberger and
+                  Amirhossein Shantia and
+                  Anselm Haselhoff},
+  title        = {Multivariate Confidence Calibration for Object Detection},
+  booktitle    = {2020 {IEEE/CVF} Conference on Computer Vision and Pattern Recognition,
+                  {CVPR} Workshops 2020, Seattle, WA, USA, June 14-19, 2020},
+  pages        = {1322--1330},
+  publisher    = {Computer Vision Foundation / {IEEE}},
+  year         = {2020},
+  url          = {https://openaccess.thecvf.com/content\_CVPRW\_2020/html/w20/Kuppers\_Multivariate\_Confidence\_Calibration\_for\_Object\_Detection\_CVPRW\_2020\_paper.html},
+  doi          = {10.1109/CVPRW50498.2020.00171},
+  timestamp    = {Sat, 30 Sep 2023 09:38:15 +0200},
+  biburl       = {https://dblp.org/rec/conf/cvpr/KuppersKSH20.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
 
 @article{Cox,
@@ -334,13 +347,23 @@
 }
 
 @inproceedings{nixon_ace,
-  title = {Measuring Calibration in Deep Learning.},
-  author = {Nixon, Jeremy and Dusenberry, Michael W and Zhang, Linchuan and Jerfel, Ghassen and Tran, Dustin},
-  booktitle = {CVPR workshops},
-  volume = {2},
-  number = {7},
-  year = {2019}
+  author       = {Jeremy Nixon and
+                  Michael W. Dusenberry and
+                  Linchuan Zhang and
[…]
