+++ paper/paper.bib
@@ -136,7 +136,8 @@
   title = {Applied logistic regression},
   author = {Hosmer Jr, David W and Lemeshow, Stanley and Sturdivant, Rodney X},
   year = {2013},
-  publisher = {John Wiley \& Sons}
+  publisher = {John Wiley \& Sons},
+  isbn = {9780470582473},
 }
 
 @article{Brocker_decompose,
@@ -216,17 +217,22 @@
 }
 
 @article{taquet2022mapie,
-  title = {MAPIE: an open-source library for distribution-free uncertainty quantification},
-  author = {Taquet, Vianney and Blot, Vincent and Morzadec, Thomas and Lacombe, Louis and Brunel, Nicolas},
-  journal = {arXiv preprint arXiv:2207.12274},
-  year = {2022}
+      title={MAPIE: an open-source library for distribution-free uncertainty quantification}, 
+      author={Vianney Taquet and Vincent Blot and Thomas Morzadec and Louis Lacombe and Nicolas Brunel},
+      year={2022},
+      eprint={2207.12274},
+      archivePrefix={arXiv},
+      primaryClass={stat.ML},
+      url={https://arxiv.org/abs/2207.12274}, 
+      doi={10.48550/arXiv.2207.12274}
 }
 
 @article{uncertaintyToolbox,
   title = {Uncertainty Toolbox: an Open-Source Library for Assessing, Visualizing, and Improving Uncertainty Quantification},
   author = {Chung, Youngseog and Char, Ian and Guo, Han and Schneider, Jeff and Neiswanger, Willie},
   journal = {arXiv preprint arXiv:2109.10254},
-  year = {2021}
+  year = {2021},
+  doi = {10.48550/arXiv.2109.10254}
 }
 
 @Manual{ResourceSelection,
@@ -248,13 +254,23 @@
   year = {2011}
 }
 
-@InProceedings{netcal,
-  author = {Küppers, Fabian and Kronenberger, Jan and Shantia, Amirhossein and Haselhoff, Anselm},
-  title = {Multivariate Confidence Calibration for Object Detection},
-  doi = {10.1109/cvprw50498.2020.00171},
-  booktitle = {The IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR) Workshops},
-  month = {June},
-  year = {2020}
+
+@inproceedings{netcal,
+  author       = {Fabian K{\"{u}}ppers and
+                  Jan Kronenberger and
+                  Amirhossein Shantia and
+                  Anselm Haselhoff},
+  title        = {Multivariate Confidence Calibration for Object Detection},
+  booktitle    = {2020 {IEEE/CVF} Conference on Computer Vision and Pattern Recognition,
+                  {CVPR} Workshops 2020, Seattle, WA, USA, June 14-19, 2020},
+  pages        = {1322--1330},
+  publisher    = {Computer Vision Foundation / {IEEE}},
+  year         = {2020},
+  url          = {https://openaccess.thecvf.com/content\_CVPRW\_2020/html/w20/Kuppers\_Multivariate\_Confidence\_Calibration\_for\_Object\_Detection\_CVPRW\_2020\_paper.html},
+  doi          = {10.1109/CVPRW50498.2020.00171},
+  timestamp    = {Sat, 30 Sep 2023 09:38:15 +0200},
+  biburl       = {https://dblp.org/rec/conf/cvpr/KuppersKSH20.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
 
 @article{Cox,
@@ -334,12 +350,21 @@
 }
 
 @inproceedings{nixon_ace,
-  title = {Measuring Calibration in Deep Learning.},
[…]
