+++ paper/paper.md
@@ -172,7 +172,7 @@ where $\eta$ is the testing data prevalence, $\eta'$ is the training data preval
 `calzone` supports multiclass classification using a 1-vs-rest approach or top-class calibration. In top-class calibration, class 1 probability is the highest predicted probability, and class 0 is 1 minus this probability. Metrics interpretation may change in this transformation.
 
 ## Verification of methods
-`calzone` results were compared with external packages for accuracy. Reliability diagrams were verified with `sklearn.calibration.calibration_curve()`[@scikit], top-class ECE and Spiegelhalter's Z scores with `MAPIE`[@taquet2022mapie], and Hosmer-Lemeshow statistic with `ResourceSelection` [@ResourceSelection] in R. Differences were within 0.1%, confirming consistency. Verification codes are in the documentation.
+To ensure the accuracy and reliability of the metrics implemented in `calzone`, we performed comprehensive validation against established external packages. Reliability diagrams were compared with `sklearn.calibration.calibration_curve()`[@scikit], top-class ECE and Spiegelhalter's Z scores were validated against `MAPIE`[@taquet2022mapie], and the Hosmer-Lemeshow statistic was checked against `ResourceSelection` [@ResourceSelection] in R. Additional tests were conducted using the `relplot` and `pycaleva` Python packages to further confirm metric consistency. All differences were within 0.1%, demonstrating strong agreement. These validation tests are documented in `test_results.py`. Furthermore, synthetic data tests (see `test_metrics.py`) were used to confirm the expected behavior of the calibration metrics under controlled conditions.
 
 ## Command line interface
 `calzone` offers a command line interface for visualizing calibration curves, calculating metrics, and confidence intervals. Run `python cal_metrics.py -h` for help.
