+++ tests/test_results.py
@@ -1,0 +1,308 @@
+"""
+This module contains validation tests for calibration metrics against external packages.
+
+The tests validate the calzone package implementation against:
+- scikit-learn's calibration_curve for reliability diagrams
+- MAPIE's ECE and Spiegelhalter's Z test implementations
+
+Based on validation.ipynb notebook.
+"""
+
+import numpy as np
+import pytest
+from calzone.utils import reliability_diagram, data_loader
+from calzone.metrics import calculate_ece_mce, spiegelhalter_z_test, hosmer_lemeshow_test
+import os
+
+
+class TestValidationAgainstExternalPackages:
+    """Test suite for validating calzone metrics against external packages."""
+    
+    @classmethod
+    def setup_class(cls):
+        """Set up test data for all tests."""
+        # Get the path to the example data relative to the test file
+        test_dir = os.path.dirname(os.path.abspath(__file__))
+        data_path = os.path.join(test_dir, "..", "example_data", "simulated_welldata.csv")
+        cls.wellcal_dataloader = data_loader(data_path=data_path)
+    
+    def test_reliability_diagram_against_sklearn(self):
+        """Test reliability diagram implementation against scikit-learn."""
+        try:
+            from sklearn.calibration import calibration_curve
+        except ImportError:
+            pytest.skip("scikit-learn not available")
+        
+        # scikit-learn implementation
+        scikit_reliability_H, scikit_confidence_H = calibration_curve(
+            self.wellcal_dataloader.labels, 
+            self.wellcal_dataloader.probs[:, 1], 
+            n_bins=15, 
+            strategy='uniform', 
+            pos_label=1
+        )
+        scikit_reliability_C, scikit_confidence_C = calibration_curve(
+            self.wellcal_dataloader.labels, 
+            self.wellcal_dataloader.probs[:, 1], 
+            n_bins=15, 
+            strategy='quantile', 
+            pos_label=1
+        )
+        
+        # calzone implementation
+        calzone_reliability_H, calzone_confidence_H, bin_edge_H, bin_count_H = reliability_diagram(
+            self.wellcal_dataloader.labels, 
+            self.wellcal_dataloader.probs, 
+            num_bins=15, 
+            class_to_plot=1, 
+            is_equal_freq=False
+        )
+        calzone_reliability_C, calzone_confidence_C, bin_edge_C, bin_count_C = reliability_diagram(
+            self.wellcal_dataloader.labels, 
+            self.wellcal_dataloader.probs, 
+            num_bins=15, 
+            class_to_plot=1, 
+            is_equal_freq=True
+        )
+        
+        # Test equal-width binning
+        reliability_diff_H = np.abs(scikit_reliability_H - calzone_reliability_H)
+        confidence_diff_H = np.abs(scikit_confidence_H - calzone_confidence_H)
+        
+        # Assert differences are very small (within numerical precision)
+        assert np.all(reliability_diff_H < 1e-4), f"Reliability differences too large: {reliability_diff_H}"
+        assert np.all(confidence_diff_H < 1e-4), f"Confidence differences too large: {confidence_diff_H}"
+        
+        # Test equal-count binning
+        reliability_diff_C = np.abs(scikit_reliability_C - calzone_reliability_C)
+        confidence_diff_C = np.abs(scikit_confidence_C - calzone_confidence_C)
+        
+        assert np.all(reliability_diff_C < 1e-4), f"Reliability differences too large: {reliability_diff_C}"
+        assert np.all(confidence_diff_C < 1e-4), f"Confidence differences too large: {confidence_diff_C}"
+    
[…]
