### 2025-08-14 f55fa42
Update contribution.md
added more specific contribution guidelines
