+++ contribution.md
@@ -1,29 +1,35 @@
-# Contribution Guide
+# Contributing to Calzone
 
-Thank you for your interest in contributing to this project!
-
-## How to Contribute
-
-1. **Fork the repository** and create your branch from `main`.
-2. **Make your changes** with clear, descriptive commit messages.
-3. **Test your changes** to ensure they work as expected.
-4. **Submit a pull request** with a detailed description of your changes.
-
-## Code Style
-
-- Follow existing code conventions.
-- Write clear, maintainable code.
-- Document your code where necessary.
-
-## Reporting Issues
-
-- Use the issue tracker for bugs and feature requests.
-- Provide steps to reproduce, expected behavior, and screenshots if possible.
-
-## Code of Conduct
-
-Please be respectful and considerate in all interactions.
+Thank you for your interest in contributing to **Calzone**! We welcome any improvements that help ensure the project remains reliable, maintainable, and of high-quality.
 
 ---
 
-We appreciate your contributions!
+##  How to Contribute
+
+1. Fork the repository and create working branches from `main`.  
+2. Open a pull request (PR) for your changes, ideally linked to an issue that describes the feature or bug.  
+3. Ensure all CI checks pass (tests, linting, formatting) before requesting review.  
+
+---
+
+##  Project Standards and Guidelines
+
+### 1. Statistical Tests & Metrics
+
+- Every new statistical test or metric must include a **unit test** under the `tests/` directory.  
+- Whenever possible, unit tests should include:
+  - **Comparisons with results from existing packages**, or  
+  - **Analytical or known-case tests** to validate correctness.
+
+### 2. New Metrics: Confidence Intervals
+
+- Any newly introduced metric must be delivered with a method to compute **confidence intervals**.  
+- Please document:
+  - The **methodology** (e.g., parametric bootstrap, analytical derivation), and  
+  - Any **assumptions or parameters** required.
+
+### 3. Code Formatting
+
+- Before committing, run **Black** to format your Python code consistently:
+  ```bash
+  black .
