+++ docs/build/html/_sources/notebooks/metrics_summary.ipynb.txt
@@ -1,207 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# Summary and guide for calzone\n",
-    "\n",
-    "We provide a summary of the calibration metrics provides by calzone, including the pros and cons of each metrics. For a more detailed explanation of each metrics and how to calculate them using calzone, please refer to the specific notebook."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 2,
-   "metadata": {
-    "nbsphinx": "hidden"
-   },
-   "outputs": [],
-   "source": [
-    "import pandas as pd\n",
-    "from IPython.display import display, HTML\n",
-    "data = {\n",
-    "    'Metrics': ['Expected calibration error<br>(ECE)', 'Maximum calibration error<br>(MCE)', 'Hosmer-Lemeshow test', \"Spiegelhalter's z test\", \"Cox's analysis\", 'Integrated calibration index<br> (ICI)'],\n",
-    "    'Description': [\n",
-    "        '<div>Using binned reliability diagram<br>(equal-width or equal-count binning),<br>sum of absolute difference, weighted by bin count.</div>',\n",
-    "        '<div>Using binned reliability diagram<br>(equal-width or equal-count binning),<br>Maximum absolute difference.</div>',\n",
-    "        '<div>Using binned reliability diagram<br>(equal-width or equal-count binning),<br>Chi-squared based test using expected and observed.</div>',\n",
-    "        '<div>Decomposition of brier score.<br>Normal distributed<br> </div>',\n",
-    "        '<div>Logistic regression of the logits<br> <br> </div>',\n",
-    "        '<div>Similar to ECE, using smooth fit (usually losse)<br>instead of binning to get<br>the calibration curve</div>'\n",
-    "    ],\n",
-    "    'Pros': [\n",
-    "        '<div>• Intuitive<br>• Easy to calculate</div>',\n",
-    "        '<div>• Intuitive<br>• Easy to calculate</div>',\n",
-    "        '<div>• Intuitive<br>• Statistical meaning</div>',\n",
-    "        '<div>• Doesn\\'t rely on binning<br>• Statistical meaning</div>',\n",
-    "        '<div>• Doesn\\'t rely on binning<br>• Hints at miscalibration type</div>',\n",
-    "        '<div>• Doesn\\'t rely on binning<br>• Capture all kind of miscalibration</div>'\n",
-    "    ],\n",
-    "    'Cons': [\n",
-    "        '<div>• Depend on binning <br>• Depend on class-by-class/top-class</div>',\n",
-    "        '<div>• Depend on binning <br>• Depend on class-by-class/top-class</div>',\n",
-    "        '<div>• Depend on binning <br>• Low power<br>• Wrong coverage</div>',\n",
-    "        '<div>• Doesn\\'t detect prevalence shift</div>',\n",
-    "        '<div>• Failed to capture some miscalibration</div>',\n",
-    "        '<div>• Depend on the choice of curve fitting<br>• Depend on fitting parameters</div>'\n",
-    "    ],\n",
-    "    'Meaning': [\n",
-    "        '<div>Average deviation from<br>true probability</div>',\n",
-    "        '<div>Maximum deviation from<br>true probability</div>',\n",
-    "        '<div>Test of<br>calibration</div>',\n",
-    "        '<div>Test of<br>calibration</div>',\n",
-    "        '<div>A logit fit to the<br>calibration curve</div>',\n",
[…]
