+++ docs/build/html/_sources/modules.rst.txt
@@ -1,7 +1,0 @@
-calzone
-=======
-
-.. toctree::
-   :maxdepth: 4
-
-   calzone
