+++ docs/build/html/_sources/calzone.rst.txt
@@ -1,37 +1,0 @@
-calzone package
-===============
-
-Submodules
-----------
-
-calzone.metrics module
-----------------------
-
-.. automodule:: calzone.metrics
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-calzone.utils module
---------------------
-
-.. automodule:: calzone.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-calzone.vis module
-------------------
-
-.. automodule:: calzone.vis
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: calzone
-   :members:
-   :undoc-members:
-   :show-inheritance:
