+++ docs/build/html/_sources/notebooks/topclass.ipynb.txt
@@ -1,121 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "## Multiclass extension\n",
-    "\n",
-    "In the previous notebooks, most metrics only consider binary classification except $\\text{ECE}_{top}$ and $\\text{MCE}_{top}$. To handle multi-class classification, the most common way is to convert it to a series of binary classification problems by using 1-vs-rest approach. Alternativly, if the user only care about whether the predicted class of each sample is having the correct predicted probability, the user transform the data to a top-class binary classification problem. In top-class problem, the class 0 probability is the $1-p_{top}$ and class 1 probability is $p_{top}$ and the label is whether true label is equal to the top-class prediction. In `calzone`, user can transform the data using `data_loader.transform_topclass()` function. Most metrics will still works but the interpretation of the metrics will be different."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 3,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from calzone.utils import data_loader,reliability_diagram\n",
-    "from calzone.vis import plot_reliability_diagram\n",
-    "from calzone.metrics import CalibrationMetrics\n",
-    "\n",
-    "dataset = data_loader('../../../example_data/simulated_welldata.csv')\n",
-    "cal_metrics = CalibrationMetrics(class_to_calculate=1)\n",
-    "### Transform to top-class \n",
-    "transformed_data = dataset.transform_topclass()"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 4,
-   "metadata": {},
[…]

[…]
