+++ .github/workflows/unit_test.yml
@@ -34,4 +34,4 @@ jobs:
     - name: Run tests
       run: |
         if [ -f tests/test_requirements.txt ]; then pip install -r tests/test_requirements.txt; fi
-        python tests/test_metrics.py
+        python tests/test_results.py
