+++ .github/workflows/unit_test.yml
@@ -31,7 +31,9 @@ jobs:
     - name: run main metrics class
       run: |
         python tests/test_metrics.py
-    - name: Run tests
+    - name: Install test dependencies
       run: |
         if [ -f tests/test_requirements.txt ]; then pip install -r tests/test_requirements.txt; fi
-        python tests/test_results.py
+    - name: Run tests
+      run: |
+        pytest tests/test_results.py
