+++ .github/workflows/unit_test.yml
@@ -36,4 +36,4 @@ jobs:
         if [ -f tests/test_requirements.txt ]; then pip install -r tests/test_requirements.txt; fi
     - name: Run tests
       run: |
-        pytest tests/test_results.py
+        python tests/test_results.py
