### 2025-08-20 f653cf8
Update paper.md
minor changes to manuscript
