### 2025-08-27 #8026 @qiancao
@fabian-s @MrShoenel @pescap

Thank you for reviewing our package! We've been making updates to both the code and manuscript. You can find our itemized responses to both reviewers here:

[MrShoenel](https://github.com/DIDSR/calzone/issues/6)
[pescap](https://github.com/DIDSR/calzone/issues/5)

Please let us know if you have additional questions or suggestions!
