### 2025-09-05 DIDSR/calzone#6 @MrShoenel
Thanks for your comments, most of the issues were addressed properly.

- About your suggestion of keeping a dictionary: I'm fine with your motivation, mine was just a suggestion.

Here is a list of things that should be improved:

- ~~Add all required packages to `requirements.txt`: Currently, `pytest`, `scikit-learn`, `MAPIE`, `relplot`, and `pycaleva` are missing. The tests require all of these.~~ OK I just realized there is another requirements.txt file in the tests-folder. I recommend you add a (sub-) section to your `readme.md` in which you outline how to run the tests (e.g., install additional requirements, what command, etc.)
- Test coverage is only at **54%** currently (I ran `PYTHONPATH=. pytest --cov=calzone --cov-report html:htmlcov` to get these numbers).
    - It seems to me that some important border-cases are missing (e.g., in the Cox regression analysis, `fix_intercept==True` is not covered).
- I suggest a slight restructuring of the tests so that simply running `pytest` will work.
    -  I suggest to remove those try/raise checks for when importing fails. Those can also be not covered meaningfully.
