### 2025-10-07 #8026 @fabian-s
Since @pescap is unresponsive, I've taken the liberty to tick off their remaining points, seems to me they were fully adressed in your reply / revision.

See https://github.com/DIDSR/calzone/issues/11 for some typos etc to fix, then we're good to go here.
